Rework the encoding of asymmetric key types
This provides more flexibility to allocate additional asymmetric key types for PAKE and PQC algorithms.

Allocate the top bit of the FAMILY field to the ASYM-TYPE field. This makes FAMILY 6 bits and ASYM-TYPE 5.

The top bit of FAMILY was previously for IMPDEF families, but support for IMPDEF key types is already provided via bit 15 in the type. So it was not used in any specification defined family values.

Defined ASYM-TYPE 0 to be a category of non-parameterized asymmetric keys, including RSA.

* Updated the ECC family and DH family definitions. Also added clarification about the inclusion of the parity bit.
* Updated all relevant sections of the encoding appendix, including the bit field graphics and ASYM-TYPE values in the tables.
* Updated the example macros in the appendix to mask FAMILY and ASYM-TYPE fields correctly.